Encapsulate config.ini inside global_config.json.
This is a change that moves towards combining cit and fit
configurations.  This will not be the final solution but allows
both cit and fit to read from a common file.

run.py will still read from config.ini and operate as it
had been.

The table is set for run_tests.py to run CIT tests by
setting the FIT_CONFIG environment variable.

Fix an issue where USER_AUTH_PORTS needed to be strings
in order to be formatted in urls.