RAC-3883 keep new infra-logging from interfering with unconverted CIT tests' output
General solution: make it so the places where infra-logging gets imported ONLY happen when running FIT tests. Changes made to do this:
* remove with-stream-monitor from setup.cfg. This keeps nose from auto-bringing in stream-monitor which in turns brings in the new logging infrastructure
* remove access to infra-logging from fit_common.py import-time. (This doesn't actually impact CIT interaction, but does keep us from double-rotating the log-dir and/or rotating them on "run_test.py --help"
* added "--with-stream-monitor" to sub-process invocation within fit-common. That means when we DO finally kick off the real nosetests via run_tests.py , the plugin is enabled and flogging comes on line
* changed tests/rackhd11/test_rackhd11_api_catalogs.py to directly grab flogging. See below for thoughts on that.
* added another random file some test generated to .gitignore.

Note that I'm a little iffy on having the test-file import flogging directly. I still think there might be an end-case for accidently double-rotating log-dir because of this. However, it's not one of the main paths (like Jenkins or the more common test-development flows).