rac-3777 verbosity ctl
This adds control of the levels to the new logging infra. This is a first step, and
is focused mainly on the controls via the nose-plugin options, but with some wiring
for (temporary) to run_test.py's -v specturm. (rac-3984 will dig into that further).

Alas, this is a fairly big change-set. The bulk of what was added can be summed up with
the following --help chunk:

  listing log settings:
    --sm-list-logging   list logging settings
    --sm-list-level=LIST_LOGGING_LEVEL
                        detail level if --list-logging is used

  setting log options:
    note: both 'logger-name's and 'handler-name's can contain wild-card
    characters to match multiple items.

    --sm-set-logger-level=('logger-name', 'level-name-or-value')
                        Set a logger's capture threshold. Loggers are
                        evaluated before handlers.
    --sm-set-handler-level=('handler-name[:logger-name]', 'level-name-or-value')
                        Set a handler's capture threshold.
    --sm-set-combo-level=('handler-name[:logger-name]', 'level-name-or-value')
                        Set a handler's capture threshold AND all loggers
                        feeding it.
    --sm-crank-file=('file-pattern', '[handler-name[:logger-name] [level-name-or-value]]')
                        changes the logging level for handler(s) on the given
                        loggers to a new value for matching files

Overview of whats going on in order to make sense of the changes :) :
In a very brief sense, when you use the "main" logger from the first example conversion (logs = flogging.get_loggers()), you
are writing to the 'test.run' logger when you do a 'logs.debug()' or 'logs.info()' call. The sequence is a like this:

There are four infra-logging loggers: infra.run, infra.data, test.run, test.dat (and technically the root one, but it itself ins't used directly for log-calls)
So, by default, there are six files: 'test_run.log', 'test_data.log', 'infra_run.log', 'infra_data.log', 'combined_all_all.log', and 'console_capture.log', placed in the directory 'log_output/run_<datetime-stamp>.d'. 'log_output/run_last.d' is linked to the most recent of these.
test_run, test_data, infra_run, and infra_data all contain DEBUG level and above output for THEIR loggers.
combined_all_all contains DEBUG level and above from all of those.
console_capture contains only INFO level and above from the four infra-loggers (and INFO and above from any other logger outside of main infra-logging ones)

So, back to the --help stuff:
    --sm-set-logger-level=('logger-name', 'level-name-or-value')
                        Set a logger's capture threshold. Loggers are
                        evaluated before handlers.
This will change the logger level. For example, '--sm-set-logger-level test.run DEBUG_5' will change the logger level from DEBUG(_0) to DEBUG_5. I.E. DEBUG_6 will still be dropped by the test.run logger.
The way things are wired, this only impacts the level of messages flowing to test_run.log and combined_all_all.log. The messages would still be dropped by the root-logger, since it is at INFO.

Note that the logger name can be wild-carded: 'test.*' or 'test*' to change both test.run and test.data, or '*.data' to crank up the volume of data dumps.

    --sm-set-handler-level=('handler-name[:logger-name]', 'level-name-or-value')
                        Set a handler's capture threshold.

But if you wanted to have more than just DEBUG flow to stdout? '--sm-set-handler-level console DEBUG'. This will actually have less impact than you would think, since
both the root logger AND the console handler have a level threshold of 'INFO'.
The actual usefull application of this option would be to lower levels '--sm-set-handler-level console WARNING'. This would me only warnings and above to the console, while still sending INFO and above to the console_capture file.

Mostly, this is here for completeness, and for use 'behind the scenes' on the more usefull next option:

    --sm-set-combo-level=('handler-name[:logger-name]', 'level-name-or-value')
                        Set a handler's capture threshold AND all loggers
                        feeding it.

Doing a '--sm-set-combo-level console DEBUG' will go and change the handler level for 'console' AND all loggers feeding into it. Because 'console' is attached to the
root logger, this ends up override ALL the loggers (that had been created at this point. That is a key limitation).
Of course, this means that the console is being spammed with DEBUG output from everything. '--sm-set-combo-level console*:*.run DEBUG_5' will have the affect of:

    --sm-crank-file=('file-pattern', '[handler-name[:logger-name] [level-name-or-value]]')
                        changes the logging level for handler(s) on the given
                        loggers to a new value for matching files

This last option can be used to "crank up" the output for a specific file or files. The handler and logger name parts are just like in the combo.
The file pattern is a regex applied against the filename seen in the output records.

The changes made to do this:
* adapted run_test.py to be able to pass arguments it didn't use to nosetests (and get help back).
* changed the comment for the -v option in run_test.py and setup argument mappings to implement them.
* changed the wrapper in _LevelLoggerClass, which was wrongly using _log rather than log (_log didn't propagate!)
* changed levels in the dict-config from DEBUG to NOTSET infra handlers
* added a 'reset_config' method in flogging/infra_logging.py to squash configs between tests run in the plugin tester
* added a 'flogging/infra_logopts.py':
** contains a parser-helper class that other entry-points can call to add the infra-logging-opts options to it and then later call to process the args. This also includes the implementation for the list and set options.
** since nose is still on the old optargs, add a converter class to convert setup via argparse to the more limited optargs.
** the _LoggingConfigFilter class defined here implements the handler+logger filtering described above.
* wired the parser-helper setup and processing into the stream_monitor plugin.
* moved most of the file observer logic into its own file, log_stream_Helper.py
* added logic to same to deal with testing for level based logs.
* added a test_log_opts.py inside stream-monitor/test to self-test options.
* rewired imports a little to be more precise about when the "real" flogging gets imported/instantiated.