rac-44440-unicode-in-test-run-arguments
This change tries to prevent the annoyingly common situation of cutting and pasting something that replaces a '-' with a unicode en-dash. The code does:
* tries to substitute certain known culprits.
* blows up if there are still unicode characters that won't convert to ascii when the substitutions are done.

There are probably more we could substitute, but this at least stops you from spending all day trying to figure out why something is subtely not working!

Removed a few imports from fit_common as well (They were not being used by something importing fit_common either!)
Added a NOQA: imported but unused for "unittest", which IS used by a ton of stuff via the fit_common import.