add retries for get_sku_nodes test case
After posting the sku, the RackHD will run sku discovery workflow and
assign the node with new sku. Previously the test case hard-codes to
wait only 3 sec. But we found this delay is not enough in some system,
and the Jenkins continuous smoke test occasionally reports this test
case failure.