RAC-4753, RAC-4752
Normalize the use of amqp with/without ssl.

Add logic that looks for file named deploy-generated.json in
the config directory.  If found, this file gets applied on
top of the configuration being generated.   This allows for a
given deployment to stamp out the ports, etc it needs without
having to make changes to stack_config.json.

Move ports into install-config context